+++ costs/src/pages/Login.tsx
@@ -1,64 +1,13 @@
-import { useEffect, useState, useContext } from "react"
-import { getAuth, signInWithEmailAndPassword, signInWithPopup, GoogleAuthProvider } from "firebase/auth";
 import MainTitleWithButton from "../components/layout/MainTitleWithButton";
 import Message from "../components/layout/Message";
-import LoginForm, { UserType } from "../components/form/LoginForm";
-import { useNavigate } from "react-router-dom";
+import LoginForm from "../components/form/LoginForm";
 import { FcGoogle } from "react-icons/fc";
 import ContentForm from "../components/form/ContentForm";
-import { SetUserContext } from "../components/contexts/Contexts";
-
-const provider = new GoogleAuthProvider();
+import useLogin from "../components/hooks/useLogin";
 
 export default function Login() {
-    const setUid = useContext(SetUserContext)
-    const navigate = useNavigate()
 
-    // Message
-    const [message, setMessage] = useState<string | undefined>()
-    let type = "error";
-
-    useEffect(() => {
-        if (message) {
-            const timer = setTimeout(() => {
-                setMessage(undefined)
-            }, 5000)
-
-            return () => clearTimeout(timer)
-        }
-    },[message])
-
-    function handleLogin(user: UserType) {
-        const auth = getAuth();
-        if (user?.email && user.password) {
-            signInWithEmailAndPassword(auth, user.email, user.password)
-            .then((userCredential) => {
-                // Signed in 
-                const user = userCredential.user;
-                setUid(user.uid)
-                navigate("/projects", {state: {message: "Login realizado com sucesso", type: "success"}})
-            })
-            .catch((error) => {
-                const errorCode = error.code;
-                const errorMessage = error.message;
-                setMessage(errorCode + errorMessage)
-            });
-        }
-    }
-
-    function handleLoginGoogle() {
-        const auth = getAuth();
-        signInWithPopup(auth, provider)
-        .then((result) => {
-            const user = result.user;
-            setUid(user.uid)
-            navigate("/projects", {state: {message: "Login realizado com sucesso", type: "success"}})
-        }).catch((error) => {
-            const errorCode = error.code;
-            const errorMessage = error.message;
-            setMessage(errorCode + errorMessage)
-        });
-    }
+    const { handleLogin, handleLoginGoogle } = useLogin()
 
     return (
         <main className="flex flex-col gap-y-10">
